Fixed data types and some bugs in the models in the Auth schema. The models in the Const schema have been made compatible with the database and constraints such as fillfactor have been applied.

+++ OnlineBookClub.WEB/Models/DB/Const/City.cs
@@ -4,10 +4,12 @@ namespace OnlineBookClub.WEB.Models.DB.Const
 {
     public class City
     {
-        public int Id { get; set; }
+        public Int16 Id { get; set; }
 
         [MaxLength(32)]
         public string Name { get; set; } = null!;
+
+        //?=========> REFERANCES
 
         public virtual List<District> Districts { get; set; }
     }
